wayland: Add cosmic-output-management
