+++ src/wayland/protocols/output_configuration/handlers/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+mod cosmic;
+mod wlr;
+
+pub use self::{cosmic::*, wlr::*};
